Makes preview  with a link from an article
index.php file contains a function which gets a full text of an article and the link to the page with the article as parametres. The function returns the variable which contains the shortened  to 180 symbols text of the article + ellipsis. If the initial article contains less than 180 symbols, the text isn't shortened and the ellipsis isn't added. But in both cases last two words are made the link leading to the page with full text of an article.